model-service/requirements.txt
# Model Service Requirements
# Python dependencies for multimodal deepfake detection system

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers for wav2vec2
transformers>=4.30.0

# Computer Vision
timm>=0.9.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# Voice Activity Detection (optional)
# pyannote-audio>=3.0.0  # Uncomment if using pyannote VAD

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Utilities
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Logging and Monitoring
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: Async job queue (for production)
# celery>=5.3.0
# redis>=4.5.0
